requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-admin-interface==0.16.3
django-chartjs==2.2.1
django-colorfield==0.4.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
